some more test cases with some fixes too
