+++ pkg/detectors/apimatic/apimatic.go
@@ -2,17 +2,17 @@ package apimatic
 
 import (
 	"context"
-	regexp "github.com/wasilibs/go-re2"
 	"net/http"
-	"strings"
 	"time"
+
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/pb/detectorspb"
 )
 
-type Scanner struct{
+type Scanner struct {
 	detectors.DefaultMultiPartCredentialProvider
 }
 
@@ -23,8 +23,7 @@ var (
 	client = common.SaneHttpClient()
 
 	// Make sure that your group is surrounded in boundary characters such as below to reduce false positives.
-	keyPat  = regexp.MustCompile(detectors.PrefixRegex([]string{"apimatic"}) + `\b([a-zA-Z0-9]{3,20}@[a-zA-Z0-9]{2,12}.[a-zA-Z0-9]{2,5})\b`)
-	passPat = regexp.MustCompile(detectors.PrefixRegex([]string{"apimatic"}) + `\b([a-z0-9-\S]{8,32})\b`)
+	apiKeyPat = regexp.MustCompile(detectors.PrefixRegex([]string{"apimatic", "apikey"}) + `\b([a-zA-Z0-9_-]{64})\b`)
 )
 
 // Keywords are used for efficiently pre-filtering chunks.
@@ -37,47 +36,43 @@ func (s Scanner) Keywords() []string {
 func (s Scanner) FromData(ctx context.Context, verify bool, data []byte) (results []detectors.Result, err error) {
 	dataStr := string(data)
 
-	matches := keyPat.FindAllStringSubmatch(dataStr, -1)
-	passMatches := passPat.FindAllStringSubmatch(dataStr, -1)
+	uniqueApiKeys := make(map[string]struct{})
+	for _, matches := range apiKeyPat.FindAllStringSubmatch(dataStr, -1) {
+		uniqueApiKeys[matches[1]] = struct{}{}
+	}
 
-	for _, match := range matches {
-		if len(match) != 2 {
-			continue
+	for apiKey := range uniqueApiKeys {
+		s1 := detectors.Result{
+			DetectorType: detectorspb.DetectorType_APIMatic,
+			Raw:          []byte(apiKey),
 		}
-		userPatMatch := strings.TrimSpace(match[1])
 
-		for _, idMatch := range passMatches {
-			if len(idMatch) != 2 {
+		if verify {
+			timeout := 10 * time.Second
+			client.Timeout = timeout
+
+			// api docs: https://docs.apimatic.io/platform-api/#/http/api-endpoints/code-generation-external-apis/list-all-code-generations
+			req, err := http.NewRequestWithContext(ctx, "GET", "https://api.apimatic.io/code-generations", http.NoBody)
+			if err != nil {
 				continue
 			}
 
-			passPatMatch := strings.TrimSpace(idMatch[1])
-
-			s1 := detectors.Result{
-				DetectorType: detectorspb.DetectorType_APIMatic,
-				Raw:          []byte(userPatMatch),
-				RawV2:        []byte(userPatMatch + passPatMatch),
-			}
-			if verify {
-				timeout := 10 * time.Second
-				client.Timeout = timeout
-				req, err := http.NewRequestWithContext(ctx, "GET", "https://www.apimatic.io/api/code-generations", nil)
-				if err != nil {
-					continue
+			// authentication documentation: https://docs.apimatic.io/platform-api/#/http/guides/authentication
+			req.Header.Set("Authorization", "X-Auth-Key "+apiKey)
+			resp, err := client.Do(req)
+			if err == nil {
+				defer resp.Body.Close()
+				if resp.StatusCode == 200 {
+					s1.Verified = true
 				}
-				req.SetBasicAuth(userPatMatch, passPatMatch)
-				res, err := client.Do(req)
-
-				if err == nil {
-					defer res.Body.Close()
-					if res.StatusCode >= 200 && res.StatusCode < 300 {
-						s1.Verified = true
-					}
-				}
+			} else {
+				// if any error happens during api request capture that as verificaiton error
+				s1.SetVerificationError(err, apiKey)
 			}
-
-			results = append(results, s1)
 		}
+
+		results = append(results, s1)
+
 	}
 	return results, nil
 }
